Add files via upload
This project involves building a time series forecasting model using a recurrent neural network (RNN) architecture implemented with the Keras library in Python. The goal is to predict future stock prices based on historical data. The project begins with fetching historical stock price data for a company, preprocessing it by scaling the values using a MinMaxScaler, and then splitting it into training and testing datasets. A deep learning model is constructed using multiple LSTM layers followed by dense layers for prediction. The model is trained on the training dataset and evaluated on the testing dataset using mean squared error as the loss function. After training, the model is used to predict future stock prices on the testing dataset. Finally, the predicted prices are compared with the actual prices to assess the model's performance. The results are visualized using Matplotlib, providing insights into how well the model predicts future stock prices. Overall, this project demonstrates the application of deep learning techniques for time series forecasting in the financial domain.